[feat] CRUD delete offercard

+++ app/api/offer/route.ts
@@ -41,14 +41,36 @@ export async function POST(req: NextRequest, res: NextResponse) {
 }
 
 export async function DELETE(req: NextRequest, res: NextResponse) {
-  const cardId = await req.text();
-  const cardIdParsed = await JSON.parse(cardId);
+  // const { id } = await req.text();
+  try {
+    const offerInfo = await req.text();
+    const offerInfoData = await JSON.parse(offerInfo);
 
-  await prisma.donationInfo.delete({
-    where: {
-      id: cardIdParsed,
-    },
-  });
+    // USE getServerSession.email to find user in DB
+    // retrieve user ID
 
-  return new Response('donation Info Deleted');
+    const session = await getServerSession(authHandler);
+
+    if (session?.user?.email) {
+      const databaseUser = await prisma.userInfo.findUnique({
+        where: {
+          personalEmail: session?.user?.email,
+        },
+      });
+
+      if (!databaseUser) {
+        return new Response('Giveaway not found', { status: 404 });
+      }
+
+      await prisma.donationInfo.delete({
+        where: {
+          id: offerInfoData,
+        },
+      });
+    }
+    return new Response('OfferCard deleted');
+  } catch (error) {
+    console.error('An error occurred while deleting the giveaway', error);
+    return new Response('Error deleting the giveaway', { status: 500 });
+  }
 }
